refactor: split the existing code into two definable components
